[raft/scd] Extract create and update subscription
